+++ src/main/java/raf/draft/dsw/controller/states/SelectState.java
@@ -5,17 +5,17 @@ import raf.draft.dsw.model.structures.room.interfaces.VisualElement;
 
 public class SelectState extends AbstractState{
     @Override
-    public void mouseClick(int x, int y, VisualElement element, RoomTab roomTab) {
+    public void mouseClick(double x, double y, VisualElement element, RoomTab roomTab) {
         // Select element
     }
 
     @Override
-    void mousePressed(int x, int y, VisualElement element, RoomTab roomTab) {
+    void mousePressed(double x, double y, VisualElement element, RoomTab roomTab) {
         // Create new selection rectangle
     }
 
     @Override
-    public void mouseDragged(int dx, int dy, VisualElement element, RoomTab roomTab) {
+    public void mouseDragged(double dx, double dy, VisualElement element, RoomTab roomTab) {
         // Scale selection rectangle
     }
 }
